resolve duplicate navbar and authentication UI issues
