+++ src/components/AdminDashboard.jsx
@@ -5,15 +5,6 @@ const AdminDashboard = () => {
   const navigate = useNavigate();
   const [issues, setIssues] = useState([]);
 
-  // Logout handler
-  const handleLogout = () => {
-    localStorage.removeItem('token'); // Remove token from localStorage
-    toast.success("you have been logged out")
-    setTimeout(() => {
-      navigate('/login');
-    }, 2000);
-  };
-
   // Fetch all reported issues
   const fetchIssues = React.useCallback(async () => {
     try {
@@ -68,19 +59,9 @@ const AdminDashboard = () => {
 
   return (
     <div className="min-h-screen bg-gray-50">
-      {/* Header */}
-      <header className="flex justify-between items-center px-6 py-4 bg-emerald-500 text-white shadow-md">
-        <h2 className="text-xl font-bold">Admin Dashboard</h2>
-        <button
-          onClick={handleLogout}
-          className="bg-white text-emerald-600 px-4 py-2 rounded-md font-medium hover:bg-gray-100 transition"
-        >
-          Logout
-        </button>
-      </header>
-
       {/* Issues Table */}
       <main className="p-6">
+        <h2 className="text-xl font-bold mb-4">Admin Dashboard</h2>
         <p>Welcome, Admin! You have access to admin controls.</p>
         <ToastContainer
         position="top-right"
